Allow subsequent calls to `.nouns` and `.adjs` to change constraints.
In the previous commit, the constraints provided to `.nouns` and `.adjs`
were only applied when those iterators were initialized. If you provided
different constraints in subsequent calls, they were ignored. Fixed
that.

Also changed the behavior of the iterator to return words skipped by
constraints or the exclusion list to the pool of available words each
time it yields a word, so that they can still be used if constraints
change or the exclusion list is modified later.